Added Railtie + dummy application harness to test
The initial release of Valhammer didn't include the magic to wire it
into Rails. Oops.
