+++ .gitignore
@@ -7,6 +7,7 @@
 /pkg/
 /spec/reports/
 /spec/db/*.sqlite3
+/spec/dummy/log/
 /log/
 /tmp/
 *.bundle
