+++ lib/valhammer.rb
@@ -4,3 +4,4 @@ module Valhammer
 end
 
 require 'valhammer/validations'
+require 'valhammer/railtie' if defined?(Rails::Railtie)
