+++ spec/dummy/app/models/capability.rb
@@ -1,0 +1,5 @@
+class Capability < ActiveRecord::Base
+  valhammer
+
+  belongs_to :organisation
+end
