+++ spec/dummy/app/models/organisation.rb
@@ -1,0 +1,6 @@
+class Organisation < ActiveRecord::Base
+  valhammer
+
+  has_many :resources
+  has_many :capabilities
+end
