+++ spec/dummy/app/models/resource.rb
@@ -1,0 +1,5 @@
+class Resource < ActiveRecord::Base
+  valhammer
+
+  belongs_to :organisation
+end
