+++ spec/dummy/config.ru
@@ -1,0 +1,2 @@
+require ::File.expand_path('../config/environment',  __FILE__)
+run Rails.application
