+++ spec/dummy/config/database.yml
@@ -1,0 +1,5 @@
+test:
+  adapter: sqlite3
+  pool: 1
+  timeout: 5000
+  database: '../db/test.sqlite3'
