+++ spec/dummy/config/environment.rb
@@ -1,0 +1,2 @@
+require File.expand_path('../application', __FILE__)
+Rails.application.initialize!
