+++ spec/dummy/config/secrets.yml
@@ -1,0 +1,2 @@
+test:
+  secret_key_base: 890f88caa9641d1845e85a49fd9aedbd30cf507f223dcb1cb6ccc53ba23731d
