+++ valhammer.gemspec
@@ -20,6 +20,7 @@ Gem::Specification.new do |spec|
 
   spec.add_dependency 'activerecord', '~> 4.1'
 
+  spec.add_development_dependency 'rails', '~> 4.1'
   spec.add_development_dependency 'bundler', '~> 1.7'
   spec.add_development_dependency 'rake', '~> 10.0'
   spec.add_development_dependency 'guard'
